+++ scri/asymptotic_bondi_data/__init__.py
@@ -1,4 +1,3 @@
-from math import sqrt, pi
 import numpy as np
 from spherical_functions import LM_total_size
 from .. import ModesTimeSeries
@@ -155,55 +154,6 @@ class AsymptoticBondiData:
 
     from .from_initial_values import from_initial_values
 
-    def mass_aspect(self, truncate_ell=max):
-        """Compute the Bondi mass aspect of the AsymptoticBondiData
-
-        The Bondi mass aspect is given by
-
-            \\Psi = \\psi_2 + \\eth \\eth \bar{\\sigma} + \\sigma * \\dot{\bar{\\sigma}}
-
-        Note that the last term is a product between two fields.  If, for example, these both have
-        ell_max=8, then their full product would have ell_max=16, meaning that we would go from
-        tracking 77 modes to 289.  This shows that deciding how to truncate the output ell is
-        important, which is why this function has the extra argument that it does.
-
-        Parameters
-        ==========
-        truncate_ell: int, or callable [defaults to `max`]
-            Determines how the ell_max value of the output is determined.  If an integer is passed,
-            each term in the output is truncated to have at most that ell_max.  (In particular,
-            terms that will not be used in the output are simply not computed, without incurring any
-            errors due to aliasing.)  If a callable is passed, it is passed on to the
-            spherical_functions.Modes.multiply method.  See that function's docstring for details.
-            The default behavior will result in the output having ell_max equal to the largest of
-            any of the individual Modes objects in the equation for \\Psi above -- but not the
-            product.
-
-        """
-        if callable(truncate_ell):
-            return self.psi2 + self.sigma.bar.eth.eth + self.sigma.multiply(self.sigma.bar.dot, truncator=truncate_ell)
-        elif truncate_ell:
-            return (
-                self.psi2.truncate_ell(truncate_ell)
-                + self.sigma.bar.eth.eth.truncate_ell(truncate_ell)
-                + self.sigma.multiply(self.sigma.bar.dot, truncator=lambda tup: truncate_ell)
-            )
-        else:
-            return self.psi2 + self.sigma.bar.eth.eth + self.sigma * self.sigma.bar.dot
-
-    @property
-    def bondi_four_momentum(self):
-        """Compute the Bondi four-momentum of the AsymptoticBondiData"""
-        import spherical_functions as sf
-
-        P_restricted = -self.mass_aspect(1).view(np.ndarray) / sqrt(4 * pi)  # Compute only the parts we need, ell<=1
-        four_momentum = np.empty(P_restricted.shape, dtype=float)
-        four_momentum[..., 0] = P_restricted[..., 0].real
-        four_momentum[..., 1] = (P_restricted[..., 3] - P_restricted[..., 1]).real / sqrt(6)
-        four_momentum[..., 2] = (1j * (P_restricted[..., 3] + P_restricted[..., 1])).real / sqrt(6)
-        four_momentum[..., 3] = -P_restricted[..., 2].real / sqrt(3)
-        return four_momentum
-
     from .constraints import (
         bondi_constraints,
         bondi_violations,
@@ -217,3 +167,4 @@ class AsymptoticBondiData:
     )
 
     from .transformations import transform
+    from .bms_charges import mass_aspect, bondi_four_momentum
